+++ extensions/porkbun/CHANGELOG.md
@@ -1,3 +1,15 @@
 # Porkbun Changelog
 
+## - {PR_MERGE_DATE}
+
+### Enhancements
+
+- Force IPv4 via Preferences
+
+### New Endpoints
+
+- Retrieve All Domains
+- Get Name Servers
+- Update Name Servers
+
 ## [Initial Version] - 2023-04-15